requirements.txt
aiohttp~=3.8.6
PyJWT~=2.8.0
httpx~=0.25.0
uvicorn~=0.23.2
fastapi~=0.103.2
pytest~=7.4.2
requests~=2.31.0
Jinja2~=3.1.2